Add an explicit cast to avoid a compiler warning
openct/proto-t1.c:775:10: warning: comparison of integers of different signs:
      'int' and 'unsigned int' [-Wsign-compare]
                        || (n != 4 + t1->rc_bytes) ...
                            ~ ^  ~~~~~~~~~~~~~~~~


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/Drivers/ccid@5516 0ce88b0d-b2fd-0310-8134-9614164e65ea

+++ src/openct/proto-t1.c
@@ -772,7 +772,7 @@ int t1_negotiate_ifsd(t1_state_t * t1, unsigned int dad, int ifsd)
 			|| (sdata[DATA] != ifsd)				/* Wrong ifsd received */
 			|| (sdata[NAD] != swap_nibbles(dad))	/* wrong NAD */
 			|| (!t1_verify_checksum(t1, sdata, n))	/* checksum failed */
-			|| (n != 4 + t1->rc_bytes)				/* wrong frame length */
+			|| (n != 4 + (int)t1->rc_bytes)				/* wrong frame length */
 			|| (sdata[LEN] != 1)					/* wrong data length */
 			|| (sdata[PCB] != (T1_S_BLOCK | T1_S_RESPONSE | T1_S_IFS))) /* wrong PCB */
 			continue;
